refactor: finish the cleanup pass on the remaining untrimmed files
Trim comments that restate the code or narrate structure, fix accumulated
slop, and split the largest file in the pass.

- nostr/subscriber.rs (817 lines) splits into subscriber/{mod,publish,query}.rs
  along a clean seam: connection + DM transport, event publishing, relay
  queries. The three groups share only `client`/`keys` and never call each
  other.
- subscribe_to_pod_events nested 8 levels deep inside the notification
  closure; per-kind decoding moves to `decode_inbound_event`, leaving a flat
  decode-then-dispatch body.
- Five send_* wrappers each repeated serialize-then-send; they now funnel
  through a private generic `send_json_private_message`.
- durable_workload.rs: a doc comment explaining why only warm-standby acts on
  quorum loss had been glued onto `fail`, where it described the wrong
  function. Moved to `failover_eligible`. Also drops a per-observation
  String clone in `tick`.
- provider/config.rs: `.context(format!(..))` allocated on the happy path in
  load_config/save_config; now `.with_context(|| ..)`.
- tests: the synthetic Cashu V3 token fixture was duplicated between
  cashu_redemption.rs and cli_deploy.rs; it moves to tests/common/mod.rs.

No behavior changes. Wire-format, event-kind, `#[serde]` and incident
comments are kept, as is clap help text, which is user-facing output rather
than commentary.
